rename FormsView to FormsSubmissionsDetail
