+++ ui/src/pages/FormsSubmissionsDetail/InputComponent.jsx
@@ -27,7 +27,7 @@ import IconCalendarMonth from '@mui/icons-material/CalendarMonth'
 import IconInsertDriveFile from '@mui/icons-material/InsertDriveFile'
 
 // STYLES
-import useStyles from './formsViewUseStyles'
+import useStyles from './formsSubmissionsDetailUseStyles'
 
 const InputComponent = (props) => {
   const { type, defaultValue } = props
