+++ ui/src/pages/FormsSubmissionsDetail/ItemGrid.jsx
@@ -4,7 +4,7 @@ import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
 
 // STYLES
-import useStyles from './formsViewUseStyles'
+import useStyles from './formsSubmissionsDetailUseStyles'
 
 const ItemGrid = (props) => {
   const { label, description, isSeparator, children } = props
